requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dateutil>=2.8
